Clase2/Ejemplo1_ConsultaBureauCrédito/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-mock==3.12.0
requests==2.31.0
httpx==0.25.2
faker==20.1.0
pydantic==2.5.2
python-dotenv==1.0.0
